requirements.txt
pandas>=2.0.0
rapidfuzz>=3.0.0
groq>=0.9.0
python-dotenv>=1.0.0
pytest>=7.0.0